No longer reading the URL parameter if the config specifies a main module.
